Additional tests in vector-vs-array
 - Attempt to incorporate Daniel Lemire's "fair" implementation
 - Attempt to use emplace_back instead of push_back
 - (un)fair implementation that uses pointer to underlying array

This version attempts to fix the initial criticisms that I received on my
first version. Most significant was the critique that I was comparing a
static data structure with a dynamic data structure (I still maintain that
my data structure did not grow with added elements, so the comparison was
"fair". As an attempt to increase speed, I tried emplace_back instead of
push_back to prevent initialisation+copy. Finally, I pit Daniel's version
against mine, using vector::resize() to resize the vector, and using
C++11's for all loop semantics to insert elements.

The last implementation is a cheap hack that attempts to use a pointer to
set all values directly, using vector::data(), bypassing the STL
specifications. However, this is not found to introduce any significant
gains as compared to the for all loop initialisation.